+++ .ci/test-project/packages/uberenv-magictestlib/package.py
@@ -1,5 +1,5 @@
 ###############################################################################
-# Copyright (c) 2014-2022, Lawrence Livermore National Security, LLC.
+# Copyright (c) 2014-2025, Lawrence Livermore National Security, LLC.
 #
 # Produced at the Lawrence Livermore National Laboratory
 #
@@ -42,13 +42,10 @@
 #
 ###############################################################################
 
-from spack import *
+from spack.package import *
 
 import socket
 import os
-
-from os.path import join as pjoin
-from os import environ as env
 
 from .magictestlib import Magictestlib
 
@@ -65,3 +62,9 @@ class UberenvMagictestlib(Magictestlib):
         dummy_tar_path = os.path.abspath(os.path.join(__file__, "../uberenv-magictestlib.tar.gz"))
         url = "file://" + dummy_tar_path
         return url
+
+    def hostconfig(self, spec, prefix):
+        super().hostconfig(spec, prefix)
+        src = self._get_host_config_path(self.spec)
+        dst = join_path(self.spec.prefix, os.path.basename(src))
+        copy(src, dst)