+++ gen_spack_env_script.py
@@ -1,5 +1,5 @@
 ###############################################################################
-# Copyright (c) 2015-2023, Lawrence Livermore National Security, LLC.
+# Copyright (c) 2015-2024, Lawrence Livermore National Security, LLC.
 #
 # Produced at the Lawrence Livermore National Laboratory
 #